[CL-3172] Add mock
