+++ front/app/components/PostShowComponents/Comments/CommentFooter.tsx
@@ -158,7 +158,6 @@ const CommentFooter = ({
         <StyledCommentVote
           postId={postId}
           postType={postType}
-          commentId={commentId}
           comment={comment}
           commentType={commentType}
         />
